refactor(1013): remove FilterOperatorEngine facade (SC-001, position 40)
Extract FilterOperatorEngine to src/utils/filter_operator_engine.py as a
pure-utility module (no MistHelper live-global dependencies). The class
provides the operator catalog (12 operators), value-required + remote-
prefilter subsets, MAC / free-text normalization helpers, and the operator
evaluator used by client search filtering.

- New module: src/utils/filter_operator_engine.py (130 LOC)
- MistHelper.py: replaced class body with breadcrumb; added re-export alias
- pyproject.toml: added coverage-omit entry for new module

Local gate: ruff/black/py_compile clean; --test 66/0/131 (baseline match).
